UI Changes

+++ Application/Models/PerformedWorks/EditPerformedWorkModel.cs
@@ -25,6 +25,6 @@ namespace Application.Models.PerformedWorks
 
         [Required]
         [Range(1, int.MaxValue)]
-        public int FuelPrice { get; init; }
+        public decimal FuelPrice { get; init; }
     }
 }
