+++ Application/Models/PerformedWorks/GetPerformedWorkModel.cs
@@ -19,6 +19,6 @@ namespace Application.Models.PerformedWorks
 
         public int AmountOfFuel { get; init; }
 
-        public int FuelPrice { get; init; }
+        public decimal FuelPrice { get; init; }
     }
 }
